requirements.txt
asgiref==3.5.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
protobuf==4.21.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.19.2
pytz==2021.3
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
